refactor: review improvements
