+++ botfront/cypress/fixtures/test-project-cypress/credentials.yml
@@ -2,3 +2,4 @@ rasa_addons.core.channels.webchat.WebchatInputTest:
   session_persistence: true
   base_url: 'http://localhost:5005'
   socket_path: /socket.io/
+rasa_addons.core.channels.bot_regression_test.BotRegressionTestInput: {}
