+++ botfront/imports/api/graphql/project/import.utils.js
@@ -80,7 +80,7 @@ const createTimestamp = () => (
 
 export async function validateFiles(files, params) {
     let filesWithMessages = files;
-    let newParams = { ...params, summary: [], timestamp: createTimestamp() };
+    let newParams = { ...params, timestamp: createTimestamp() };
     // this is the validation pipeline each step only add errors to the files it should validate
     // each step can also add data to the params, eg : the default domain, the summary of changes etc,
     [filesWithMessages, newParams] = validateDefaultDomains(filesWithMessages, newParams);
